All files necessary for the MCT exercise.
Included is the index file for the local node server, the accompanying package.json for building the project, and the three static files in the public directory. index.html includes the minimal markup and links to the main.js file including all the client-side javascript necessary and a small styles.css file.